Solved initialization bug
The subcatchments children were deactivated in random order due to the concurrencies between OnEnable and Awake functions. This created errors with the display and sound of the subcatchment's infrastructures that leverage the OnEnable function.

Side note: github says that some things were changed in the SampleScene.unity file, this probably depends on me opening the project from unity hub and not directly from file. If after this commit there are problems with cloned projects, take the commits before this as a possible starting point for debugging.